Version 1.11
Improving keyboard functionality:  Added possibility to control slide completely with keyboard, which is useful for imaging experiments. With this responseKeys were added. One can change the number of pixels the slide moves each time with stepSize (Default is 1).

In addition a bug with slidercolor  was fixed and abortime's default was changed from 8 seconds to Inf.